info: Change server MD5 to SHA1
This commit replaces the Minecraft 1.8.8 server.jar MD5 hash with the
SHA1 hash. This change helps us use the official launcher download link
in KigPaper now that the Amazon AWS link was shut down.

Please note that this is not a security feature. While being
significantly stronger than MD5, SHA1 should not be considered a secure
hashing algorithm nowadays.